# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dienstplan-App API für Persönliche Assistenz
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: Schichten auflisten
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [active, standby, night, full_day, vacation, sick, work]
      responses:
        "200":
          description: Liste der Schichten
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Schicht anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftInput"
      responses:
        "201":
          description: Erstellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    get:
      operationId: getShift
      tags: [shifts]
      summary: Schicht abrufen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schicht
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "404":
          description: Nicht gefunden
    patch:
      operationId: updateShift
      tags: [shifts]
      summary: Schicht aktualisieren
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftUpdate"
      responses:
        "200":
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Schicht löschen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Gelöscht
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: Teams auflisten
      responses:
        "200":
          description: Liste der Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Team anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Erstellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Team aktualisieren
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Nicht gefunden
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Team löschen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Gelöscht
        "404":
          description: Nicht gefunden
        "409":
          description: Team enthält noch Daten oder Mitglieder
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, accountType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, assistant]
        accountType:
          type: string
          enum: [privat, dienstleister]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        socialSecurityNumber:
          type: ["string", "null"]
        taxId:
          type: ["string", "null"]
        taxClass:
          type: ["string", "null"]
        healthInsurance:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Team:
      type: object
      required: [id, name, ownerId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        ownerId:
          type: integer
        createdAt:
          type: string
          format: date-time
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    TeamUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
...
    Shift:
      type: object
      required: [id, userId, startTime, endTime, type, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        type:
          type: string
          enum: [active, standby, night, full_day, vacation, sick, work]
        shiftModelId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        valuedHours:
          type: number
        nightHours:
          type: number
        sundayHours:
          type: number
        holidayHours:
          type: number
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    ShiftInput:
      type: object
      required: [userId, startTime, endTime, type]
      properties:
        userId:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        type:
          type: string
          enum: [active, standby, night, full_day, vacation, sick, work]
        shiftModelId:
          type: ["integer", "null"]
        notes:
          type: string
    ShiftUpdate:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        type:
          type: string
          enum: [active, standby, night, full_day, vacation, sick, work]
        shiftModelId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
...