# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dienstplan-App API für Persönliche Assistenz
 * OpenAPI spec version: 0.1.0
 */

export interface TeamUpdate {
  /** @minLength 1 */
  name: string;
}